refactor: rename about and contacts dir and classes to main-content and about-us
